Refactor queen tracker out of backfiller
